chore: update Quotas ModelSnapshot

+++ Modules/Quotas/src/Quotas.Infrastructure.Database.SqlServer/Migrations/QuotasDbContextModelSnapshot.cs
@@ -27,8 +27,8 @@ namespace Quotas.Infrastructure.Database.SqlServer.Migrations
                     b.Property<string>("Id")
                         .HasColumnType("nvarchar(450)");
 
-                    b.Property<int>("CipherSize")
-                        .HasColumnType("int");
+                    b.Property<long>("CipherSize")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("datetime2");
